feature: Backup Vault PSQL Flex Instance Restore script
The script uses the Azure REST api to trigger a restore to a storage account container.

You can pass in the RECOVERY_POINT_TIME as "latest" and it will pull the latest by date, otherwise it needs the recoveryPointTime property in proper timestamp format.

The BACKUP_INSTANCE_NAME can be the friendly name (what the portal shows) it will reach out and get the proper id based on that.

RESOURCE_GROUP, SUBSCRIPTION_ID, refer to the VAULT_NAME, properties.

RESTORE_LOCATION needs to be the azure region, not sure if its of the storage or the vault.

Note as of this commit az rest does not have a header return so using --debug 2>debug_request.txt is the only way to get the Azure-AsyncOperation URL.